Finalised the turning fractions method in TrafficSim. What a pain. --------------------------------------------------- This method computes all turning fraction per node / tClick / incoming x outgoing link It reflects what the original matlab code does. But there could be problems with uninitialised values in the turningFractions structure.
Note that travelTimeManager class was also created.
It includes methods to compute traveltime given a travel cost for a link
and node 2 node.

ODPairs was also modified. Methods to obtain the index for a node id,
startNode and endNode where added. This is not the best place to do
this.But improving would require reverse engineering further the matlab
code, which is left for a second phase. Note that probably the solution
comes through the construction a priori of some kind of
hashmap<node.id,index>

Other classes where fixed to call the methods in travelTimeManager