Most of the LTM is done -- adding the node model is left --------------------------------------------------------
Simulator specific changes:
-Major changes 
done to cumulatives -- moved update cumulatives to the
LTM
-Created init(), implemented most of run(), updatedCumulatives in LTM
and
made a minor chage to TrafficLink in order to store the link
upstream / downstream cumulative maxima.
-Major improvement compared to matlab code max(1,t-1) has been replaced
with the maintenance of maximum values at link level, which are updated
at the end of each timeClick in run()

Missing:
-call in LTM.ru()n updateCumu
-call in updateCumu the node Model

Generic changes:
-Added package utils and created a class Pair (mainly to exchange
pairs of values between methods).
Originally
I
implemented
a very simple Pair but found a very thourough implementation in stack
overflow... That's the Pair I added in the end.